+++ src/main/java/service/command/ConvertCommand.java
@@ -45,6 +45,6 @@ public class ConvertCommand extends BaseCommand {
                 .map(inputFile -> AsyncTaskExecutor.submit(() -> converter.ncm2Mp3(inputFile.getAbsolutePath(), finalOutputPath.getAbsolutePath())))
                 .collect(Collectors.toList());
         Utils.waitForAllTask(futures, result -> result);
-        return;
+        System.exit(0);
     }
 }